Changed iterator using rect, added descriptions
coordinate exporter app, which allows you to get coordinates of structures in defined range. It does not load or generate map, so it could be very 'light'. It is slower than external amidst, but it is more precise than amidst, especially when you tries to find end cities, or end ships. Since this app does not offer any copy-paste or save options by itself, you need clients that provides logging to copy and paste, to use in other external apps.