added bucketstack.sc file
Added script to make any bucket act like water bottles except for the customizable stack size. Buckets now stack in a player's inventory, the stack remains when a bucket is used, and the player gets the empty bucket or drops it on the ground. Incompatible with Tweakeroo's tweakHandRestock. Yes, axolotl bucket has already been added, and no, it doesn't cause problems.